Minor improvements
Added better file upload security, orphan files automatic deletion and upload rate limiting. Changed user id type to Long

+++ backend/src/main/java/com/smartentrance/backend/config/RevocationLoader.java
@@ -30,7 +30,7 @@ public class RevocationLoader implements CommandLineRunner {
         List<RevocationEntry> entries = revocationRepository.findAll();
 
         if (!entries.isEmpty()) {
-            Map<Integer, Long> loadedData = new HashMap<>();
+            Map<Long, Long> loadedData = new HashMap<>();
 
             for (RevocationEntry entry : entries) {
                 loadedData.put(entry.getUserId(), entry.getRevokedAt());
