+++ backend/src/main/java/com/smartentrance/backend/repository/BuildingRepository.java
@@ -17,9 +17,9 @@ public interface BuildingRepository extends JpaRepository<Building, Integer> {
 
     boolean existsByGooglePlaceIdAndEntrance(String googlePlaceId, String entrance);
 
-    boolean existsByIdAndManagerId(Integer id, Integer managerId);
+    boolean existsByIdAndManagerId(Integer id, Long managerId);
 
-    List<Building> findAllByManagerId(Integer managerId);
+    List<Building> findAllByManagerId(Long managerId);
 
     @QueryHints(value = @QueryHint(name = HINT_FETCH_SIZE, value = "50"))
     @Query("SELECT b FROM Building b")
