+++ backend/src/main/java/com/smartentrance/backend/repository/UnitRepository.java
@@ -14,9 +14,9 @@ public interface UnitRepository extends JpaRepository<Unit, Long> {
 
     boolean existsByAccessCode(String accessCode);
 
-    List<Unit> findAllByResponsibleUserId(Integer userId);
+    List<Unit> findAllByResponsibleUserId(Long userId);
 
-    boolean existsByBuildingIdAndResponsibleUserId(Integer buildingId, Integer userId);
+    boolean existsByBuildingIdAndResponsibleUserId(Integer buildingId, Long userId);
 
     List<Unit> findAllByBuildingIdOrderByUnitNumberAsc(Integer buildingId);
